+++ app/src/main/java/app/cashadvisor/common/utill/exceptions/RegisterException.kt
@@ -39,8 +39,8 @@ sealed class RegisterException(
 
         class UnauthorizedWrongConfirmationCode(
             override val message: String,
-            val remainingAttempts: Int,
-            val lockDuration: Int,
+            val remainingAttempts: Long,
+            val lockDuration: Long,
             val statusCode: Int
         ) : EmailCodeConfirmation(message)
